move code into package structure
