+++ quiz_app/static/js/freelancer.js
